+++ Nudge-Post-install.bash
@@ -5,14 +5,18 @@
 #    Nudge Post-install
 #
 #    Purpose: Configures Nudge to company standards post-install
-#    https://github.com/macadmins/nudge/blob/main/README.md#configuration
+#    https://github.com/dan-snelson/Nudge-Post-install/wiki
 #
 ####################################################################################################
 #
-#   Version 0.0.15, 08-Jun-2022, Dan K. Snelson (@dan-snelson)
-#       Added an `Uninstall` option to the `resetConfiguration` function
-#       Removed verbosity when removing files
-#       Started macOS Ventura logic (using macOS Monterey's deadline)
+# Version 0.0.16, 21-Oct-2022, Dan K. Snelson (@dan-snelson)
+#   **BREAKING CHANGES**
+#       - Reordered Script Parameters
+#       - Added default values (for when Script Parameters are left blank in a Jamf Pro policy)
+#   Removed `authorizationCheck` function
+#   Added macOS Ventura settings
+#   Replaced `scriptResult` with `updateScriptLog`
+#   Random clean-up
 #
 ####################################################################################################
 
@@ -24,29 +28,51 @@
 #
 ####################################################################################################
 
-scriptVersion="0.0.15"
-scriptResult=""
-loggedInUser=$( /bin/echo "show State:/Users/ConsoleUser" | /usr/sbin/scutil | /usr/bin/awk '/Name :/ { print $3 }' )
-loggedInUserID=$( /usr/bin/id -u "${loggedInUser}" )
-authorizationKey="${4}"                     # Authorization Key to prevent unauthorized execution via Jamf Remote
-plistDomain="${5}"                          # Reverse Domain Name Notation (i.e., "org.churchofjesuschrist")
-resetConfiguration="${6}"                   # Configuration Files to Reset (i.e., None (blank) | All | JSON | LaunchAgent | LaunchDaemon)
-requiredBigSurMinimumOSVersion="${7}"       # Required macOS Big Sur Minimum Version (i.e., 11.6.5)
-requiredBigSurInstallationDate="${8}"       # Required macOS Big SurInstallation Date & Time (i.e., 2022-03-21T10:00:00Z)
-requiredMontereyMinimumOSVersion="${9}"     # Required macOS Monterey Minimum Version (i.e., 12.3)
-requiredMontereyInstallationDate="${10}"    # Required macOS Monterey Installation Date & Time (i.e., 2022-03-21T10:00:00Z)
-requiredVenturaMinimumOSVersion="${11}"     # Required macOS Ventura Minimum Version (i.e., 13.1)
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# Global Variables
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+scriptVersion="0.0.16"
+export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin/
+loggedInUser=$( echo "show State:/Users/ConsoleUser" | scutil | awk '/Name :/ { print $3 }' )
+plistDomain="${4:-"org.churchofjesuschrist"}"                       # Reverse Domain Name Notation (i.e., "org.churchofjesuschrist")
+resetConfiguration="${5:-"All"}"                                    # Configuration Files to Reset (i.e., None (blank) | All | JSON | LaunchAgent | LaunchDaemon)
+requiredBigSurMinimumOSVersion="${6:-"11.99"}"                      # Required macOS Big Sur Minimum Version (i.e., 11.7.1)
+requiredBigSurInstallationDate="${7:-"2022-11-24T10:00:00Z"}"       # Required macOS Big SurInstallation Date & Time (i.e., 2022-11-24T10:00:00Z)
+requiredMontereyMinimumOSVersion="${8:-"12.99"}"                    # Required macOS Monterey Minimum Version (i.e., 12.6.1)
+requiredMontereyInstallationDate="${9:-"2022-11-24T10:00:00Z"}"     # Required macOS Monterey Installation Date & Time (i.e., 2022-11-24T10:00:00Z)
+requiredVenturaMinimumOSVersion="${10:-"13.99"}"                    # Required macOS Ventura Minimum Version (i.e., 13.1)
+requiredVenturaInstallationDate="${11:-"2022-11-24T10:00:00Z"}"     # Required macOS Ventura Installation Date & Time (i.e., 2022-11-24T10:00:00Z)
+scriptLog="/var/log/${plistDomain}.log"
 jsonPath="/Library/Preferences/${plistDomain}.Nudge.json"
 launchAgentPath="/Library/LaunchAgents/${plistDomain}.Nudge.plist"
 launchDaemonPath="/Library/LaunchDaemons/${plistDomain}.Nudge.logger.plist"
 
+
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # Set deadline variable based on OS version
-osProductVersion=$( /usr/bin/sw_vers -productVersion )
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+osProductVersion=$( sw_vers -productVersion )
 case "${osProductVersion}" in
     11* ) deadline="${requiredBigSurInstallationDate}"    ;;
     12* ) deadline="${requiredMontereyInstallationDate}"  ;;
-    13* ) deadline="${requiredMontereyInstallationDate}"  ;;
+    13* ) deadline="${requiredVenturaInstallationDate}"   ;;
 esac
+
+
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# Validate logged-in user
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+if [[ -z "${loggedInUser}" || "${loggedInUser}" == "loginwindow" ]]; then
+    updateScriptLog "No user logged-in; exiting."
+    exit 1
+else
+    loggedInUserID=$(id -u "${loggedInUser}")
+fi
 
 
 
@@ -57,22 +83,23 @@ esac
 ####################################################################################################
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Check for a specified value in Parameter 4 to prevent unauthorized script execution
+# Client-side Script Logging
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-function authorizationCheck() {
+function updateScriptLog() {
+    echo -e "$( date +%Y-%m-%d\ %H:%M:%S ) - ${1}" | tee -a "${scriptLog}"
+}
 
-    if [[ "${authorizationKey}" != "PurpleMonkeyDishwasher" ]]; then
 
-        scriptResult+="Error: Incorrect Authorization Key; exiting."
-        echo "${scriptResult}"
-        exit 1
 
-    else
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# Run command as logged-in user (thanks, @scriptingosx!)
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-        scriptResult+="Correct Authorization Key, proceeding; "
+function runAsUser() {
 
-    fi
+    updateScriptLog "Run \"$@\" as \"$loggedInUserID\" … "
+    launchctl asuser "$loggedInUserID" sudo -u "$loggedInUser" "$@"
 
 }
 
@@ -84,13 +111,13 @@ function authorizationCheck() {
 
 function resetConfiguration() {
 
-    scriptResult+="Reset Configuration: ${1}; "
+    updateScriptLog "Reset Configuration: ${1}"
 
     case ${1} in
 
         "All" )
             # Reset JSON, LaunchAgent, LaunchDaemon, Hide Nudge
-            scriptResult+="Reset All Configuration Files … "
+            updateScriptLog "Reset All Configuration Files … "
 
             # Reset User Preferences
             # For testing only; see:
@@ -99,101 +126,100 @@ function resetConfiguration() {
 
             # echo "Reset User Preferences"
             # /bin/rm -f /Users/"${loggedInUser}"/Library/Preferences/com.github.macadmins.Nudge.plist
-            # /usr/bin/pkill -l -U "${loggedInUser}" cfprefsd
-            # scriptResult+="Removed User Preferences; "
+            # pkill -l -U "${loggedInUser}" cfprefsd
+            # updateScriptLog "Removed User Preferences"
 
             # Reset JSON
-            scriptResult+="Remove ${jsonPath} … "
+            updateScriptLog "Remove ${jsonPath} … "
             /bin/rm -f "${jsonPath}" 2>&1
-            scriptResult+="Removed ${jsonPath}; "
+            updateScriptLog "Removed ${jsonPath}"
 
             # Reset LaunchAgent
-            scriptResult+="Unload ${launchAgentPath} … "
-            /bin/launchctl asuser "${loggedInUserID}" /bin/launchctl unload -w "${launchAgentPath}" 2>&1
-            scriptResult+="Remove ${launchAgentPath} … "
-            /bin/rm -f "${launchAgentPath}"  2>&1
-            scriptResult+="Removed ${launchAgentPath}; "
+            updateScriptLog "Unload ${launchAgentPath} … "
+            runAsUser launchctl unload -w "${launchAgentPath}" 2>&1
+            updateScriptLog "Remove ${launchAgentPath} … "
+            /bin/rm -f "${launchAgentPath}" 2>&1
+            updateScriptLog "Removed ${launchAgentPath}"
 
             # Reset LaunchDaemon
-            scriptResult+="Unload ${launchDaemonPath} … "
+            updateScriptLog "Unload ${launchDaemonPath} … "
             /bin/launchctl unload -w "${launchDaemonPath}" 2>&1
-            scriptResult+="Remove ${launchDaemonPath} … "
+            updateScriptLog "Remove ${launchDaemonPath} … "
             /bin/rm -f "${launchDaemonPath}" 2>&1
-            scriptResult+="Removed ${launchDaemonPath}; "
+            updateScriptLog "Removed ${launchDaemonPath}"
 
             # Hide Nudge in Finder
-            scriptResult+="Hide Nudge in Finder … "
-            /usr/bin/chflags hidden "/Applications/Utilities/Nudge.app" 
-            scriptResult+="Hid Nudge in Finder; "
+            updateScriptLog "Hide Nudge in Finder … "
+            chflags hidden "/Applications/Utilities/Nudge.app" 
+            updateScriptLog "Hid Nudge in Finder"
 
             # Hide Nudge in Launchpad
-            scriptResult+="Hide Nudge in Launchpad … "
+            updateScriptLog "Hide Nudge in Launchpad … "
             if [[ -z "$loggedInUser" ]]; then
-                scriptResult+="Did not detect logged-in user"
+                updateScriptLog "Did not detect logged-in user"
             else
-                /usr/bin/sqlite3 $(/usr/bin/sudo find /private/var/folders \( -name com.apple.dock.launchpad -a -user "${loggedInUser}" \) 2> /dev/null)/db/db "DELETE FROM apps WHERE title='Nudge';"
-                /usr/bin/killall Dock
-                scriptResult+="Hid Nudge in Launchpad for ${loggedInUser}; "
+                sqlite3 $(sudo find /private/var/folders \( -name com.apple.dock.launchpad -a -user ${loggedInUser} \) 2> /dev/null)/db/db "DELETE FROM apps WHERE title='Nudge';"
+                killall Dock
+                updateScriptLog "Hid Nudge in Launchpad for ${loggedInUser}"
             fi
 
-            scriptResult+="Reset All Configuration Files; "
+            updateScriptLog "Reset All Configuration Files"
             ;;
 
         "Uninstall" )
            # Uninstall Nudge Post-install
-            scriptResult+="Uninstalling Nudge Post-install … "
+            updateScriptLog "Uninstalling Nudge Post-install … "
 
             # Uninstall JSON
-            /bin/rm -f "${jsonPath}"
-            scriptResult+="Uninstalled ${jsonPath}; "
+            rm -f "${jsonPath}"
+            updateScriptLog "Uninstalled ${jsonPath}"
 
             # Uninstall LaunchAgent
-            scriptResult+="Unload ${launchAgentPath} … "
-            /bin/launchctl asuser "${loggedInUserID}" /bin/launchctl unload -w "${launchAgentPath}"
-            /bin/rm -f "${launchAgentPath}"
-            scriptResult+="Uninstalled ${launchAgentPath}; "
+            updateScriptLog "Unload ${launchAgentPath} … "
+            runAsUser launchctl unload -w "${launchAgentPath}"
+            rm -f "${launchAgentPath}"
+            updateScriptLog "Uninstalled ${launchAgentPath}"
 
             # Uninstall LaunchDaemon
-            scriptResult+="Unload ${launchDaemonPath} … "
-            /bin/launchctl unload -w "${launchDaemonPath}"
-            /bin/rm -f "${launchDaemonPath}"
-            scriptResult+="Uninstalled ${launchDaemonPath}; "
+            updateScriptLog "Unload ${launchDaemonPath} … "
+            launchctl unload -w "${launchDaemonPath}"
+            rm -f "${launchDaemonPath}"
+            updateScriptLog "Uninstalled ${launchDaemonPath}"
 
             # Exit
-            scriptResult+="Uninstalled all Nudge Post-install configuration files; "
-            scriptResult+="Thanks for using Nudge Post-install!"
-            echo "${scriptResult}"
+            updateScriptLog "Uninstalled all Nudge Post-install configuration files"
+            updateScriptLog "Thanks for using Nudge Post-install!"
             exit 0
             ;;
 
         "JSON" )
             # Reset JSON
-            scriptResult+="Remove ${jsonPath} … "
-            /bin/rm -f "${jsonPath}"
-            scriptResult+="Removed ${jsonPath}; "
+            updateScriptLog "Remove ${jsonPath} … "
+            rm -f "${jsonPath}"
+            updateScriptLog "Removed ${jsonPath}"
             ;;
 
         "LaunchAgent" )
             # Reset LaunchAgent
-            scriptResult+="Unload ${launchAgentPath} … "
-            /bin/launchctl asuser "${loggedInUserID}" /bin/launchctl unload -w "${launchAgentPath}"
-            scriptResult+="Remove ${launchAgentPath} … "
-            /bin/rm -f "${launchAgentPath}"
-            scriptResult+="Removed ${launchAgentPath}; "
+            updateScriptLog "Unload ${launchAgentPath} … "
+            runAsUser launchctl unload -w "${launchAgentPath}"
+            updateScriptLog "Remove ${launchAgentPath} … "
+            rm -f "${launchAgentPath}"
+            updateScriptLog "Removed ${launchAgentPath}"
             ;;
 
         "LaunchDaemon" )
             # Reset LaunchDaemon
-            scriptResult+="Unload ${launchDaemonPath} … "
-            /bin/launchctl unload -w "${launchDaemonPath}"
-            scriptResult+="Remove ${launchDaemonPath} … "
-            /bin/rm -f "${launchDaemonPath}"
-            scriptResult+="Removed ${launchDaemonPath}; "
+            updateScriptLog "Unload ${launchDaemonPath} … "
+            launchctl unload -w "${launchDaemonPath}"
+            updateScriptLog "Remove ${launchDaemonPath} … "
+            rm -f "${launchDaemonPath}"
+            updateScriptLog "Removed ${launchDaemonPath}"
             ;;
 
         * )
             # None of the expected options was entered; don't reset anything
-            scriptResult+="None of the expected reset options was entered; don't reset anything; "
+            updateScriptLog "None of the expected reset options was entered; don't reset anything"
             ;;
 
     esac
@@ -209,18 +235,21 @@ function resetConfiguration() {
 ####################################################################################################
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# Client-side Logging
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+if [[ ! -f "${scriptLog}" ]]; then
+    touch "${scriptLog}"
+    updateScriptLog "*** Created log file ***"
+fi
+
+
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # Logging preamble
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-scriptResult+="Nudge Post-install (${scriptVersion}) "
-
-
-
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Confirm Authorization
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-
-authorizationCheck
+updateScriptLog "Nudge Post-install (${scriptVersion})"
 
 
 
@@ -232,14 +261,13 @@ resetConfiguration "${resetConfiguration}"
 
 
 
-
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # Nudge Logger LaunchDaemon
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 if [[ ! -f ${launchDaemonPath} ]]; then
 
-    scriptResult+="Create ${launchDaemonPath} … "
+    updateScriptLog "Create ${launchDaemonPath} … "
 
     cat <<EOF > "${launchDaemonPath}"
 <?xml version="1.0" encoding="UTF-8"?>
@@ -267,11 +295,11 @@ if [[ ! -f ${launchDaemonPath} ]]; then
 </plist>
 EOF
 
-    /bin/launchctl load -w "${launchDaemonPath}"
+    launchctl load -w "${launchDaemonPath}"
 
 else
 
-    scriptResult+="${launchDaemonPath} exists; "
+    updateScriptLog "${launchDaemonPath} exists"
 
 fi
 
@@ -283,11 +311,11 @@ fi
 
 if [[ ! -f ${jsonPath} ]]; then
 
-    scriptResult+="Create ${jsonPath} … "
-    /usr/bin/touch "${jsonPath}"
-    scriptResult+="Created ${jsonPath}; "
+    updateScriptLog "Create ${jsonPath} … "
+    touch "${jsonPath}"
+    updateScriptLog "Created ${jsonPath}"
 
-    scriptResult+="Write ${jsonPath} … "
+    updateScriptLog "Write ${jsonPath} … "
 
     cat <<EOF > "${jsonPath}"
 {
@@ -345,11 +373,11 @@ if [[ ! -f ${jsonPath} ]]; then
         "aboutUpdateURLs": [
             {
             "_language": "en",
-            "aboutUpdateURL": "https://www.apple.com/macos/macos-ventura-preview/"
+            "aboutUpdateURL": "https://www.apple.com/macos/ventura/"
             }
         ],
-        "majorUpgradeAppPath": "/Applications/Install macOS 13 beta.app",
-        "requiredInstallationDate": "${requiredMontereyInstallationDate}",
+        "majorUpgradeAppPath": "/Applications/Install macOS Ventura.app",
+        "requiredInstallationDate": "${requiredVenturaInstallationDate}",
         "requiredMinimumOSVersion": "${requiredVenturaMinimumOSVersion}",
         "targetedOSVersionsRule": "13"
         }
@@ -408,11 +436,11 @@ if [[ ! -f ${jsonPath} ]]; then
 }
 EOF
 
-    scriptResult+="Wrote Nudge JSON file to ${jsonPath}; "
+    updateScriptLog "Wrote Nudge JSON file to ${jsonPath}"
 
 else
 
-    scriptResult+="${jsonPath} exists; "
+    updateScriptLog "${jsonPath} exists"
 
 fi
 
@@ -424,7 +452,7 @@ fi
 
 if [[ ! -f ${launchAgentPath} ]]; then
 
-    scriptResult+="Create ${launchAgentPath} … "
+    updateScriptLog "Create ${launchAgentPath} … "
 
     cat <<EOF > "${launchAgentPath}"
 <?xml version="1.0" encoding="UTF-8"?>
@@ -460,18 +488,18 @@ if [[ ! -f ${launchAgentPath} ]]; then
 </plist>
 EOF
 
-    scriptResult+="Created ${launchAgentPath}; "
+    updateScriptLog "Created ${launchAgentPath}"
 
-    scriptResult+="Set ${launchAgentPath} file permissions ..."
+    updateScriptLog "Set ${launchAgentPath} file permissions ..."
     /usr/sbin/chown root:wheel "${launchAgentPath}"
-    /bin/chmod 644 "${launchAgentPath}"
-    /bin/chmod +x "${launchAgentPath}"
-    scriptResult+="Set ${launchAgentPath} file permissions; "
+    chmod 644 "${launchAgentPath}"
+    chmod +x "${launchAgentPath}"
+    updateScriptLog "Set ${launchAgentPath} file permissions"
 
 else
 
-    scriptResult+="${launchAgentPath} exists"
-    scriptResult+="${launchAgentPath} exists; "
+    updateScriptLog "${launchAgentPath} exists"
+    updateScriptLog "${launchAgentPath} exists"
 
 fi
 
@@ -484,20 +512,20 @@ fi
 # https://github.com/macadmins/nudge/blob/main/build_assets/postinstall-launchagent
 # Only enable the LaunchAgent if there is a user logged in, otherwise rely on built in LaunchAgent behavior
 if [[ -z "$loggedInUser" ]]; then
-    scriptResult+="Did not detect user"
+    updateScriptLog "Did not detect user"
 elif [[ "$loggedInUser" == "loginwindow" ]]; then
-    scriptResult+="Detected Loginwindow Environment"
+    updateScriptLog "Detected Loginwindow Environment"
 elif [[ "$loggedInUser" == "_mbsetupuser" ]]; then
-    scriptResult+="Detect SetupAssistant Environment"
+    updateScriptLog "Detect SetupAssistant Environment"
 elif [[ "$loggedInUser" == "root" ]]; then
-    scriptResult+="Detect root as currently logged-in user"
+    updateScriptLog "Detect root as currently logged-in user"
 else
     # Unload the LaunchAgent so it can be triggered on re-install
-    /bin/launchctl asuser "${loggedInUserID}" /bin/launchctl unload -w "${launchAgentPath}"
+    runAsUser launchctl unload -w "${launchAgentPath}"
     # Kill Nudge just in case (say someone manually opens it and not launched via LaunchAgent
-    /usr/bin/killall Nudge
+    killall Nudge
     # Load the LaunchAgent
-    /bin/launchctl asuser "${loggedInUserID}" /bin/launchctl load -w "${launchAgentPath}"
+    runAsUser launchctl load -w "${launchAgentPath}"
 fi
 
 
@@ -506,8 +534,6 @@ fi
 # Exit
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-scriptResult+="Goodbye!"
-
-echo "${scriptResult}"
+updateScriptLog "Goodbye!"
 
 exit 0